backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
httpx>=0.27.0
beautifulsoup4>=4.12.0
lxml>=5.2.0
python-dotenv>=1.0.0
google-cloud-texttospeech>=2.16.0
google-auth>=2.29.0
openai>=1.23.0
pydub>=0.25.1
slowapi>=0.1.9
python-multipart>=0.0.9
aiofiles>=23.2.1
